backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-barcode
pillow
reportlab
python-multipart